+++ packages/app-stellar/tests/02.getPublicKeys/index.ts
@@ -1,0 +1,102 @@
+import { MockDeviceConnection } from '@cypherock/sdk-interfaces';
+import { afterEach, beforeEach, describe, expect, test } from '@jest/globals';
+
+import * as StellarSdk from 'stellar-sdk';
+import { clearMocks, expectMockCalls, setupMocks } from './__helpers__';
+import fixtures from './__fixtures__';
+
+import { setStellarLib, StellarApp } from '../../src';
+
+describe('stellarApp.getPublicKeys', () => {
+  let connection: MockDeviceConnection;
+  let stellarApp: StellarApp;
+
+  beforeEach(async () => {
+    clearMocks();
+
+    setStellarLib(StellarSdk);
+    connection = await MockDeviceConnection.create();
+    stellarApp = await StellarApp.create(connection);
+  });
+
+  afterEach(async () => {
+    await stellarApp.destroy();
+  });
+
+  describe('should be able to get public keys', () => {
+    fixtures.valid.forEach(testCase => {
+      test(testCase.name, async () => {
+        const onEvent = setupMocks(testCase);
+
+        const output = await stellarApp.getPublicKeys({
+          ...testCase.params,
+          onEvent,
+        });
+        expect(output).toEqual(testCase.output);
+
+        expectMockCalls(testCase);
+      });
+    });
+  });
+
+  describe('should throw error with invalid arguments', () => {
+    fixtures.invalidArgs.forEach(testCase => {
+      test(testCase.name, async () => {
+        setupMocks(testCase);
+
+        const rejectedPromise = stellarApp.getPublicKeys(testCase.params);
+
+        await expect(rejectedPromise).rejects.toThrow(testCase.errorInstance);
+        if (testCase.errorMessage) {
+          try {
+            await rejectedPromise;
+          } catch (error: any) {
+            expect(error.message).toMatch(testCase.errorMessage);
+          }
+        }
+      });
+    });
+  });
+
+  describe('should throw error when device returns invalid data', () => {
+    fixtures.invalidData.forEach(testCase => {
+      test(testCase.name, async () => {
+        setupMocks(testCase);
+
+        const rejectedPromise = stellarApp.getPublicKeys(testCase.params);
+
+        await expect(rejectedPromise).rejects.toThrow(testCase.errorInstance);
+        if (testCase.errorMessage) {
+          try {
+            await rejectedPromise;
+          } catch (error: any) {
+            expect(error.message).toMatch(testCase.errorMessage);
+          }
+        }
+
+        expectMockCalls(testCase);
+      });
+    });
+  });
+
+  describe('should throw error when device returns error', () => {
+    fixtures.error.forEach(testCase => {
+      test(testCase.name, async () => {
+        setupMocks(testCase);
+
+        const rejectedPromise = stellarApp.getPublicKeys(testCase.params);
+
+        await expect(rejectedPromise).rejects.toThrow(testCase.errorInstance);
+        if (testCase.errorMessage) {
+          try {
+            await rejectedPromise;
+          } catch (error: any) {
+            expect(error.message).toMatch(testCase.errorMessage);
+          }
+        }
+
+        expectMockCalls(testCase);
+      });
+    });
+  });
+});