+++ packages/app-stellar/tests/03.signTxn/__fixtures__/error.ts
@@ -1,0 +1,89 @@
+import {
+  DeviceAppError,
+  DeviceAppErrorType,
+  deviceAppErrorTypeDetails,
+} from '@cypherock/sdk-interfaces';
+import { ISignTxnTestCase } from './types';
+import { queryToUint8Array, resultToUint8Array } from '../__helpers__';
+import { ISignTxnParams } from '../../../src';
+
+const sampleStellarXDR =
+  'AAAAAgAAAADCqTIZAnYwoZ3oQWx9HISjQ+docmusJBWB7UuwyB1l1gAAAGQAAVpoAAAAFQAAAAEAAAAAAAAAAAAAAAAAAAAAAAAAAQAAABxIZWxsb29vbyBCb3JpbmcgV29ybGxkZGQuLi4gAAAAAQAAAAAAAAABAAAAAJZDpSknyjuaOzYpuvE7OxpsA8/mnsNCikb6I1McklSWAAAAAAAAAAA7msoAAAAAAAAAAAA=';
+
+const commonParams: {
+  params: ISignTxnParams;
+  queries: {
+    name: string;
+    data: Uint8Array;
+  }[];
+} = {
+  params: {
+    walletId: new Uint8Array([
+      199, 89, 252, 26, 32, 135, 183, 211, 90, 220, 38, 17, 160, 103, 233, 62,
+      110, 172, 92, 20, 35, 250, 190, 146, 62, 8, 53, 86, 128, 26, 3, 187, 121,
+      64,
+    ]),
+    derivationPath: [0x80000000 + 44, 0x80000000 + 148, 0x80000000 + 0],
+    txn: {
+      xdr: sampleStellarXDR,
+      networkPassphrase: 'Test SDF Network ; September 2015',
+    },
+  },
+  queries: [
+    {
+      name: 'Initate query',
+      data: queryToUint8Array({
+        signTxn: {
+          initiate: {
+            walletId: new Uint8Array([
+              199, 89, 252, 26, 32, 135, 183, 211, 90, 220, 38, 17, 160, 103,
+              233, 62, 110, 172, 92, 20, 35, 250, 190, 146, 62, 8, 53, 86, 128,
+              26, 3, 187, 121, 64,
+            ]),
+            derivationPath: [0x80000000 + 44, 0x80000000 + 148, 0x80000000 + 0], // CORRECTED
+            transactionSize: Buffer.from(sampleStellarXDR, 'base64').length,
+          },
+        },
+      }),
+    },
+  ],
+};
+
+const withUnknownError: ISignTxnTestCase = {
+  name: 'With unknown error',
+  ...commonParams,
+  results: [
+    {
+      name: 'error',
+      data: resultToUint8Array({
+        signTxn: {
+          commonError: {
+            unknownError: 1,
+          },
+        },
+      }),
+    },
+  ],
+  errorInstance: DeviceAppError,
+  errorMessage:
+    deviceAppErrorTypeDetails[DeviceAppErrorType.UNKNOWN_ERROR].message,
+};
+
+const withInvalidResult: ISignTxnTestCase = {
+  name: 'With invalid result',
+  ...commonParams,
+  results: [
+    {
+      name: 'error',
+      data: new Uint8Array([18, 4, 26, 2, 24, 1]),
+    },
+  ],
+  errorInstance: DeviceAppError,
+  errorMessage:
+    deviceAppErrorTypeDetails[DeviceAppErrorType.INVALID_MSG_FROM_DEVICE]
+      .message,
+};
+
+const error: ISignTxnTestCase[] = [withUnknownError, withInvalidResult];
+
+export default error;