+++ packages/app-stellar/tests/03.signTxn/__fixtures__/invalidData.ts
@@ -1,0 +1,113 @@
+import {
+  DeviceAppError,
+  DeviceAppErrorType,
+  deviceAppErrorTypeDetails,
+} from '@cypherock/sdk-interfaces';
+import { ISignTxnTestCase } from './types';
+import { queryToUint8Array } from '../__helpers__';
+import { ISignTxnParams } from '../../../src';
+
+const sampleStellarXDR =
+  'AAAAAgAAAADCqTIZAnYwoZ3oQWx9HISjQ+docmusJBWB7UuwyB1l1gAAAGQAAVpoAAAAFQAAAAEAAAAAAAAAAAAAAAAAAAAAAAAAAQAAABxIZWxsb29vbyBCb3JpbmcgV29ybGxkZGQuLi4gAAAAAQAAAAAAAAABAAAAAJZDpSknyjuaOzYpuvE7OxpsA8/mnsNCikb6I1McklSWAAAAAAAAAAA7msoAAAAAAAAAAAA=';
+
+const commonParams: {
+  params: ISignTxnParams;
+  queries: {
+    name: string;
+    data: Uint8Array;
+  }[];
+  errorInstance?: any;
+  [key: string]: any;
+} = {
+  params: {
+    walletId: new Uint8Array([
+      199, 89, 252, 26, 32, 135, 183, 211, 90, 220, 38, 17, 160, 103, 233, 62,
+      110, 172, 92, 20, 35, 250, 190, 146, 62, 8, 53, 86, 128, 26, 3, 187, 121,
+      64,
+    ]),
+    derivationPath: [0x80000000 + 44, 0x80000000 + 148, 0x80000000 + 0],
+    txn: {
+      xdr: sampleStellarXDR,
+      networkPassphrase: 'Test SDF Network ; September 2015',
+    },
+  },
+  queries: [
+    {
+      name: 'Initate query',
+      data: queryToUint8Array({
+        signTxn: {
+          initiate: {
+            walletId: new Uint8Array([
+              199, 89, 252, 26, 32, 135, 183, 211, 90, 220, 38, 17, 160, 103,
+              233, 62, 110, 172, 92, 20, 35, 250, 190, 146, 62, 8, 53, 86, 128,
+              26, 3, 187, 121, 64,
+            ]),
+            derivationPath: [0x80000000 + 44, 0x80000000 + 148, 0x80000000 + 0],
+            transactionSize: Buffer.from(sampleStellarXDR, 'base64').length,
+          },
+        },
+      }),
+    },
+  ],
+  errorInstance: DeviceAppError,
+  errorMessage:
+    deviceAppErrorTypeDetails[DeviceAppErrorType.INVALID_MSG_FROM_DEVICE]
+      .message,
+};
+
+const invalidData: ISignTxnTestCase[] = [
+  {
+    name: 'Invalid data',
+    ...commonParams,
+    results: [
+      {
+        name: 'error',
+        data: new Uint8Array([
+          109, 112, 102, 98, 72, 57, 117, 109, 75, 69, 83, 117, 117, 49, 103,
+          78, 100, 105, 87, 83, 116, 106, 71, 54, 67, 110, 104, 77, 86, 49, 113,
+          97, 78, 111, 50, 98, 118, 52, 67, 113, 72, 122, 120, 85, 98, 53, 86,
+          68, 115, 86, 52, 77, 86, 112, 83, 70, 86, 78, 121, 121, 109, 83, 112,
+          98, 74, 76, 55, 57, 75, 89, 86, 57, 75, 56, 88, 82, 100, 105, 98, 70,
+          109, 118, 54, 116, 86, 54, 116, 50, 122, 52, 100, 87, 110, 111, 110,
+          78, 52, 78, 77, 89, 109,
+        ]),
+      },
+    ],
+  },
+  {
+    name: 'Invalid data',
+    ...commonParams,
+    results: [
+      {
+        name: 'error',
+        data: new Uint8Array([
+          10, 34, 10, 3, 90, 221, 135, 18, 2, 8, 1, 24, 1, 34, 11, 8, 2, 18, 7,
+          8,
+        ]),
+      },
+    ],
+  },
+  {
+    name: 'Invalid data',
+    ...commonParams,
+
+    results: [
+      {
+        name: 'error',
+        data: new Uint8Array([10]),
+      },
+    ],
+  },
+  {
+    name: 'Invalid data',
+    ...commonParams,
+    results: [
+      {
+        name: 'error',
+        data: new Uint8Array([]),
+      },
+    ],
+  },
+];
+
+export default invalidData;